1、自定义按钮支持执行自定义JS脚本 2、修复EruptOperateLog在oracle数据源无法自动创建的问题 3、更新前端源码

+++ erupt-security/src/main/java/xyz/erupt/security/interceptor/EruptSecurityInterceptor.java
@@ -182,7 +182,7 @@ public class EruptSecurityInterceptor implements AsyncHandlerInterceptor {
                         operate.setReqParam(findRequestParamVal(request));
                     }
                     operate.setCreateTime(new Date());
-                    operate.setTotalTime((int) (operate.getCreateTime().getTime() - RequestBodyTL.get().getDate()));
+                    operate.setTotalTime(operate.getCreateTime().getTime() - RequestBodyTL.get().getDate());
                     RequestBodyTL.remove();
                     Optional.ofNullable(ex).ifPresent(e -> {
                         operate.setErrorInfo(ExceptionUtils.getStackTrace(e));
